# src/dotnet/App/Models/Base64UploadTransmission.cs
/*
 * SOCOMAP
 *
 * This API is for the new Socomap Protocol
 *
 * OpenAPI spec version: 0.0.1
 * Contact: development@infotech.de
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace Org.OpenAPITools.Models
{ 
    /// <summary>
    /// json version with base64 encoding
    /// </summary>
    [DataContract]
    public partial class Base64UploadTransmission : IEquatable<Base64UploadTransmission>
    { 
        /// <summary>
        ///  if tne client doesnt support the application/octet-stream format, the data can be uploaded as json. 
        /// </summary>
        /// <value> if tne client doesnt support the application/octet-stream format, the data can be uploaded as json. </value>
        [DataMember(Name="message")]
        public string Message { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class Base64UploadTransmission {\n");
            sb.Append("  Message: ").Append(Message).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (obj is null) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((Base64UploadTransmission)obj);
        }

        /// <summary>
        /// Returns true if Base64UploadTransmission instances are equal
        /// </summary>
        /// <param name="other">Instance of Base64UploadTransmission to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(Base64UploadTransmission other)
        {
            if (other is null) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    Message == other.Message ||
                    Message != null &&
                    Message.Equals(other.Message)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (Message != null)
                    hashCode = hashCode * 59 + Message.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(Base64UploadTransmission left, Base64UploadTransmission right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(Base64UploadTransmission left, Base64UploadTransmission right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}